Visualization refactoring
